+++ server/ASC.ZoomService/Controllers/ZoomController.cs
@@ -25,7 +25,6 @@
 // International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
 
 using ASC.ApiSystem.Helpers;
-using ASC.ApiSystem.Services.NotifyService;
 using ASC.Core.Common.Quota;
 using ASC.Core.Common.Quota.Features;
 using ASC.FederatedLogin;
@@ -36,7 +35,6 @@ using ASC.Files.Core.ApiModels.ResponseDto;
 using ASC.Files.Core.VirtualRooms;
 using ASC.Web.Api.Core;
 using ASC.Web.Core.Files;
-using ASC.Web.Core.Notify;
 using ASC.Web.Files.Classes;
 using ASC.Web.Files.Services.WCFService;
 using ASC.Web.Files.Utils;
@@ -69,6 +67,7 @@ public class ZoomController : ControllerBase
     private TenantManager TenantManager { get; }
     private CoreSettings CoreSettings { get; }
     private ZoomAccountHelper ZoomAccountHelper { get; }
+    private ZoomMultiTenantHelper ZoomMultiTenantHelper { get; }
     private TimeZonesProvider TimeZonesProvider { get; }
     private TimeZoneConverter TimeZoneConverter { get; }
     private UserManager UserManager { get; }
@@ -94,6 +93,7 @@ public class ZoomController : ControllerBase
         TenantManager tenantManager,
         CoreSettings coreSettings,
         ZoomAccountHelper zoomAccountHelper,
+        ZoomMultiTenantHelper zoomMultiTenantHelper,
         TimeZonesProvider timeZonesProvider,
         TimeZoneConverter timeZoneConverter,
         UserManager userManager,
@@ -118,6 +118,7 @@ public class ZoomController : ControllerBase
         Configuration = configuration;
         TenantManager = tenantManager;
         CoreSettings = coreSettings;
+        ZoomMultiTenantHelper = zoomMultiTenantHelper;
         ZoomAccountHelper = zoomAccountHelper;
         TimeZonesProvider = timeZonesProvider;
         TimeZoneConverter = timeZoneConverter;
@@ -159,7 +160,7 @@ public class ZoomController : ControllerBase
     [HttpGet("state")]
     [AllowCrossSiteJson]
     [Authorize(AuthenticationSchemes = ZoomAuthHandler.ZOOM_AUTH_SCHEME_HEADER)]
-    public async Task<IActionResult> GetState([FromQuery] ZoomStateModel model, [FromQuery] bool noRedirect = false)
+    public async Task<IActionResult> GetState([FromQuery] ZoomStateModel model, [FromQuery] bool noRedirect = false, [FromQuery] bool forceAuth = false)
     {
         var uid = User.Claims.FirstOrDefault(c => c.Type == ZoomAuthHandler.ZOOM_CLAIM_UID)?.Value;
         var mid = User.Claims.FirstOrDefault(c => c.Type == ZoomAuthHandler.ZOOM_CLAIM_MID)?.Value;
@@ -179,24 +180,48 @@ public class ZoomController : ControllerBase
             collaboration = Cache.GetCollaboration(mid);
         }
 
-        string confirmLink;
+        var jwtSecret = Configuration["zoom:gate-secret"];
+
+        string confirmLink = null;
         bool foreignTenant = false;
+        int tenantId = -1;
         if (collaboration != null)
         {
             Log.LogDebug($"GetState(): Collaboration is not null, getting confirm link using tenant id {collaboration.TenantId}");
             confirmLink = await GetConfirmLinkByTenantId(collaboration.TenantId, uid);
             model.TenantId = collaboration.TenantId;
+            tenantId = collaboration.TenantId;
 
             var ownTenant = GetTenantByAccountId(model.AccountId);
             foreignTenant = ownTenant.Id != collaboration.TenantId;
         }
         else
         {
-            Log.LogDebug($"GetState(): Collaboration is null, getting confirm link using account number {model.AccountId}");
-            confirmLink = await GetConfirmLinkByAccountId(model.AccountId, uid);
+            if (Request.Cookies.TryGetValue("ZoomLink", out var zoomLinkJwt))
+            {
+                try
+                {
+                    Log.LogDebug($"GetState(): ZoomLink Cookie is not null, getting confirm link using jwt");
+                    var zoomLinkJson = JsonWebToken.Decode(zoomLinkJwt, jwtSecret);
+                    var zoomLink = JsonSerializer.Deserialize<ZoomLinkCookie>(zoomLinkJson, new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
+                    confirmLink = await GetConfirmLinkByTenantId(zoomLink.TenantId, uid);
+                    tenantId = zoomLink.TenantId;
+                }
+                catch (Exception ex)
+                {
+                    Log.LogWarning(ex, $"GetState(): ZoomLink jwt could not be parsed");
+                    Response.Cookies.Delete("ZoomLink");
+                }
+            }
+
+            if (confirmLink == null)
+            {
+                Log.LogDebug($"GetState(): Collaboration is null, getting confirm link using account number {model.AccountId}");
+                confirmLink = await GetConfirmLinkByAccountId(model.AccountId, uid);
+            }
         }
 
-        if (confirmLink != null)
+        if (!forceAuth && confirmLink != null)
         {
             Log.LogDebug($"GetState(): Got request from Zoom App; Found portal and user, redirecting with auth; AccountNumber: {model.AccountId}; UserId: {uid}");
 
@@ -225,8 +250,8 @@ public class ZoomController : ControllerBase
                 }
             }
 
-            var link = collaboration != null
-                    ? GetPayloadRedirectLinkByTenantId(collaboration.TenantId, integrationPayload)
+            var link = tenantId > -1
+                    ? await GetPayloadRedirectLinkByTenantId(tenantId, integrationPayload)
                     : await GetPayloadRedirectLinkByAccountId(model.AccountId, integrationPayload);
 
             if (noRedirect)
@@ -241,7 +266,6 @@ public class ZoomController : ControllerBase
 
         Log.LogDebug($"GetState(): ConfirmLink is null, proceeding to oauth");
 
-        var jwtSecret = Configuration["zoom:gate-secret"];
 
         var verifier = GenerateCodeVerifier();
         var challenge = GenerateCodeChallenge(verifier);
@@ -318,6 +342,95 @@ public class ZoomController : ControllerBase
         }
     }
 
+    [HttpPost("link")]
+    public async Task<IActionResult> PostLink(ZoomLinkPayload model)
+    {
+        List<Tenant> tenants;
+        if (!string.IsNullOrWhiteSpace(model.Login) && !string.IsNullOrWhiteSpace(model.Password))
+        {
+            tenants = await ZoomMultiTenantHelper.FindTenantsAsync(model.Login, model.Password);
+        }
+        else
+        {
+            tenants = [];
+        }
+
+        var response = new ZoomLinkResponse()
+        {
+            Login = model.Login,
+            TenantInfo = tenants.Select(t => new ZoomTenantInfo() { Id = t.Id, Name = t.Name, Domain = t.GetTenantDomain(CoreSettings) }).ToList()
+        };
+
+        var jwtSecret = Configuration["zoom:gate-secret"];
+
+        return Ok(JsonWebToken.Encode(response, jwtSecret));
+    }
+
+    [HttpPut("link")]
+    public async Task<IActionResult> PutLink(ZoomLinkPutPayload model)
+    {
+        var response = new ZoomIntegrationPayload()
+        {
+            Home = Configuration["zoom:home"]
+        };
+
+        try
+        {
+            var jwtSecret = Configuration["zoom:gate-secret"];
+
+            var stateJson = JsonWebToken.Decode(model.State, jwtSecret);
+            var state = JsonSerializer.Deserialize<ZoomLinkResponse>(stateJson, new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
+
+            var codeVerifier = Cache.GetOauthVerifier(model.Challenge);
+            if (codeVerifier == null)
+            {
+                Log.LogDebug("PutLink(): Incorrect ouath state");
+                return BadRequest("incorrect ouath state");
+            }
+
+            if (!state.TenantInfo.Any(t => t.Id == model.ChosenTenant))
+            {
+                return BadRequest("incorrect chosen tenant");
+            }
+
+            Log.LogDebug($"PutLink(): Got PUT request from Zoom App; TenantId: {model.ChosenTenant}; User: {state.Login}");
+
+            var loginProvider = ZoomAccountHelper.GetLoginProvider();
+            Log.LogDebug("PutLink(): Exchanging code for AccessToken");
+            var token = loginProvider.GetAccessToken(model.Code, model.RedirectUri, codeVerifier);
+            Log.LogDebug("PutLink(): Requesting profile info");
+            var profile = loginProvider.GetLoginProfile(token);
+
+            Log.LogDebug("PutLink(): Creating user and/or tenant");
+            var tenant = await LinkUserToTenant(profile, state.Login, model.ChosenTenant);
+
+            Log.LogDebug($"PutLink(): Setting csp settings to allow '{$"https://{tenant.Alias}.{Configuration["zoom:zoom-domain"]}"}'.");
+            await AddDomainToCsp($"https://{tenant.Alias}.{Configuration["zoom:zoom-domain"]}");
+
+            response.ConfirmLink = GetTenantRedirectUri(tenant, state.Login);
+            response.Collaboration = new ZoomCollaborationRoom()
+            {
+                Status = ZoomCollaborationStatus.Pending,
+            };
+
+            var cookie = new ZoomLinkCookie()
+            {
+                TenantId = tenant.Id,
+            };
+
+            Response.Cookies.Append("ZoomLink", JsonWebToken.Encode(cookie, jwtSecret), new CookieOptions() { Domain = Configuration["zoom:zoom-domain"], Expires = DateTimeOffset.Now.AddDays(30) });
+
+            Log.LogDebug("PutLink(): Returning user with confirmLink");
+            return Ok(GetPayloadRedirectLink(tenant, response));
+        }
+        catch (Exception ex)
+        {
+            Log.LogDebug(ex, $"PutLink(): Error while linking");
+            response.Error = ex.Message;
+            return BadRequest(response);
+        }
+    }
+
     [HttpPost("home")]
     [AllowCrossSiteJson]
     public async Task<IActionResult> PostHome(ZoomHomeModel model)
@@ -347,10 +460,16 @@ public class ZoomController : ControllerBase
             Log.LogDebug("PostHome(): Exchanging code for AccessToken");
             var token = loginProvider.GetAccessToken(model.Code, model.RedirectUri, codeVerifier);
             Log.LogDebug("PostHome(): Requesting profile info");
-            var profile = loginProvider.GetLoginProfile(token);
+            var (profile, raw) = loginProvider.GetLoginProfileAndRaw(token.AccessToken);
+
+            if (raw.AccountId.Contains('_'))
+            {
+                // ToDo: we might have a collision
+                raw.AccountId = raw.AccountId.Replace("_", "--");
+            }
 
             Log.LogDebug("PostHome(): Creating user and/or tenant");
-            var (_, tenant) = await CreateUserAndTenant(profile, state.AccountId, state.TenantId);
+            var (_, tenant) = await CreateUserAndTenant(profile, raw.AccountId, state.TenantId);
 
             response.ConfirmLink = GetTenantRedirectUri(tenant, profile.EMail);
             if (!string.IsNullOrWhiteSpace(state.CollaborationId) && !"none".Equals(state.CollaborationId))
@@ -363,6 +482,7 @@ public class ZoomController : ControllerBase
             }
 
             Log.LogDebug("PostHome(): Returning user with confirmLink");
+            Response.Cookies.Delete("ZoomLink", new CookieOptions() { Domain = Configuration["zoom:zoom-domain"], Expires = DateTimeOffset.MinValue });
             return Ok(GetPayloadRedirectLink(tenant, response));
         }
         catch (Exception ex)
@@ -428,14 +548,14 @@ public class ZoomController : ControllerBase
             await SecurityContext.AuthenticateMeWithoutCookieAsync(Core.Configuration.Constants.CoreSystem);
             Log.LogInformation($"DeauthorizationHook(): Got deauth request with zoom user id {zoomEvent.Payload.UserId}");
             // getting all linked accounts on all tenants
-            var userIds = await AccountLinker.GetLinkedObjectsAsync(zoomEvent.Payload.UserId, ProviderConstants.Zoom);
+            var userIds = await AccountLinker.GetLinkedProfilesAsync(zoomEvent.Payload.UserId, ProviderConstants.Zoom);
 
             foreach (var userId in userIds)
             {
                 try
                 {
                     Log.LogInformation($"DeauthorizationHook(): Unlinking user with zoom id {zoomEvent.Payload.UserId}, user id {userId}");
-                    await AccountLinker.RemoveLinkAsync(userId.ToString(), zoomEvent.Payload.UserId, ProviderConstants.Zoom);
+                    await AccountLinker.RemoveProviderAsync(userId.ToString(), ProviderConstants.Zoom);
                 }
                 catch (Exception ex)
                 {
@@ -473,9 +593,9 @@ public class ZoomController : ControllerBase
         return 10 * 1024 * 1024;
     }
 
-    private string GetPayloadRedirectLinkByTenantId(int tenantId, ZoomIntegrationPayload payload)
+    private async Task<string> GetPayloadRedirectLinkByTenantId(int tenantId, ZoomIntegrationPayload payload)
     {
-        var tenant = HostedSolution.GetTenant(tenantId);
+        var tenant = await HostedSolution.GetTenantAsync(tenantId);
         return GetPayloadRedirectLink(tenant, payload);
     }
 
@@ -524,7 +644,7 @@ public class ZoomController : ControllerBase
             {
                 Log.LogDebug($"CreateUserAndTenant(): TenantId not equals accountNumber, adding user as a guest");
                 guest = true;
-                tenant = HostedSolution.GetTenant(tenantId.Value);
+                tenant = await HostedSolution.GetTenantAsync(tenantId.Value);
             }
         }
 
@@ -546,9 +666,9 @@ public class ZoomController : ControllerBase
             var userInfo = await UserManager.GetUserByEmailAsync(profile.EMail);
             if (!UserManager.UserExists(userInfo.Id))
             {
-                Log.LogDebug($"CreateUserAndTenant(): Creating new user for portal '{portalName}'; UserId: {profile.UniqueId}");
+                Log.LogDebug($"CreateUserAndTenant(): Creating new user for portal '{portalName}'; UserId: {profile.HashId}");
                 userInfo = await CreateUser(profile, guest);
-                Log.LogInformation($"CreateUserAndTenant(): Created new user for '{portalName}'; UserId: {profile.UniqueId}");
+                Log.LogInformation($"CreateUserAndTenant(): Created new user for '{portalName}'; UserId: {profile.HashId}");
                 shouldLink = true;
             }
             else
@@ -560,11 +680,58 @@ public class ZoomController : ControllerBase
             if (shouldLink)
             {
                 Log.LogDebug($"CreateUserAndTenant(): Linking portal user '{userInfo.Id}' to zoom user '{profile.Id}'.");
-                await AccountLinker.AddLinkAsync(userInfo.Id.ToString(), profile);
+
+                var links = await AccountLinker.GetLinkedProfilesAsync(userInfo.Id.ToString(), ProviderConstants.Zoom);
+                if (links.Any())
+                {
+                    Log.LogInformation($"CreateUserAndTenant(): Portal user '{userInfo.Id}' already has zoom link.");
+                    throw new Exception("User already linked");
+                }
+
+                await AccountLinker.AddLinkAsync(userInfo.Id, profile);
                 Log.LogInformation($"CreateUserAndTenant(): Linked portal user '{userInfo.Id}' to zoom user '{profile.Id}'.");
             }
 
             return (userInfo, tenant);
+        }
+        catch (Exception ex)
+        {
+            Log.LogDebug(ex, $"CreateUserAndTenant(): Error");
+            throw;
+        }
+        finally
+        {
+            SecurityContext.Logout();
+        }
+    }
+
+    private async Task<Tenant> LinkUserToTenant(LoginProfile profile, string email, int tenantId)
+    {
+        var tenant = await HostedSolution.GetTenantAsync(tenantId);
+        TenantManager.SetCurrentTenant(tenant);
+        try
+        {
+            await SecurityContext.AuthenticateMeWithoutCookieAsync(Core.Configuration.Constants.CoreSystem);
+
+            var linkedUserId = await ZoomAccountHelper.GetUserIdFromZoomUid(profile.Id);
+            if (linkedUserId != null)
+            {
+                Log.LogInformation($"LinkUserToTenant(): User already linked.");
+                return tenant;
+            }
+
+            var userInfo = await UserManager.GetUserByEmailAsync(email);
+            if (!UserManager.UserExists(userInfo.Id))
+            {
+                Log.LogInformation($"LinkUserToTenant(): User not found with email '{email}', tenant '{tenantId}'");
+                throw new Exception("User not found");
+            }
+
+            Log.LogDebug($"LinkUserToTenant(): Linking portal user '{userInfo.Id}' to zoom user '{profile.Id}'.");
+            await AccountLinker.AddLinkAsync(userInfo.Id, profile);
+            Log.LogInformation($"LinkUserToTenant(): Linked portal user '{userInfo.Id}' to zoom user '{profile.Id}'.");
+
+            return tenant;
         }
         finally
         {
@@ -691,8 +858,7 @@ public class ZoomController : ControllerBase
             }
 
             Log.LogDebug($"CreateTenant(): Setting csp settings to allow '{$"https://{portalName}.{Configuration["zoom:zoom-domain"]}"}'.");
-
-            await CspSettingsHelper.SaveAsync(new List<string>() { $"https://{portalName}.{Configuration["zoom:zoom-domain"]}" }, false);
+            await AddDomainToCsp($"https://{portalName}.{Configuration["zoom:zoom-domain"]}");
         }
         catch (Exception ex)
         {
@@ -716,6 +882,12 @@ public class ZoomController : ControllerBase
         {
             Log.LogError(ex, "Error while sending welcome email");
         }
+    }
+
+    private async Task AddDomainToCsp(string domain)
+    {
+        var domains = await CspSettingsHelper.LoadAsync();
+        await CspSettingsHelper.SaveAsync(new List<string>(domains.Domains) { domain });
     }
 
     private bool TryGetQuotaId(out int quotaId)
@@ -755,7 +927,7 @@ public class ZoomController : ControllerBase
     {
         ArgumentException.ThrowIfNullOrEmpty(uid, nameof(uid));
 
-        var tenant = HostedSolution.GetTenant(tenantId);
+        var tenant = await HostedSolution.GetTenantAsync(tenantId);
 
         Log.LogDebug($"GetConfirmLinkByTenantId(): Getting confirm link with tenant {tenant?.Id}, user {uid}.");
         return await GetConfirmLink(tenant, uid);